requirements.txt
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
greenlet==3.2.2
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
pillow==11.2.1
psycopg2==2.9.10
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.0
python-multipart==0.0.20
qrcode==8.2
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.13.2
uvicorn==0.34.2
python-jose==3.3.0

